additional_resources/smia_kb/src/requirements.txt
connexion == 2.14.2
connexion[swagger-ui] >= 2.6.0
python-dateutil == 2.9.0.post0
setuptools == 65.5.0
swagger-ui-bundle == 0.0.9
owlready2 == 0.47
